Rename classes
Use keystore-instance verification
